task2: updates to part2
These are the important changes that happened from task1 to task2:

There is a new pointer to the node structure called 'middle'. It will store the middle of the list everytime a new value is added. It is very useful for binary search.

There is a dynamic array structure, used to store the same values of the linked list. It is used to make comparisions between both structures.
Its type is int64_t, which will be used to store two 32-bit integers: the first 32-bit refers to the value and the last 32-bit refers how manytimes this value was repeated in the list.

New methods created:

int search(int32_t) : it will search the given value in the list and return its position. Or return -1, if the value was not found.

void ini_array() : it will initialize the dynamic array with the values from the linked list. One of the most interesting/important things about this method, is the part where the bit-shiffiting is done.

int32_t get_high(int64_t) : it will get the "HIGH part" of the 64-bit integer. (The value)

int32_t get_low(int64_t) : it will get the "LOW part" of the 64-bit integer. (The count of how many times the value was repeated in the list)

int32_t my_bin_search(int32_t value, void *data, int type) : this method implements the binary search in the linked list or dynamic array.
About the paramters:
value - the value that will be searched
*data - linked list or dynamic array
type - 0(linked list) or 1(dynamic array)

The implementation of the binary search for the linked list was a little bit different from the conventional mode. Everytime the search is "divided by half", the algorithm will count how many "steps" the element must go forward or backwards and then, use the pointers 'prev' and 'next' to do this movement. (It's a little bit complicated to explain, but it works)

This function also measures the time of execution, which is used to compare the exectuion time between the 2 structures: linked list and dynamic memory.

void set_middle() : it will run everytime a new value is added in order to define the middle value of the list. This method can be optmized by using the current middle value to define the new middle value, instead of going through the list every time.

void get_user_search_input() : this method will get the values the user want to search in the list. It will search the value using the method my_bin_search() in the two structures: linked list and dynamic array. The reason to run this method for the two structures is to compare the execution time between both structures.

void show_size_of_structures() : this method will show the size of both structures. At this point, I don't know if it is calculating the size correctly.

Important observation: the list was implemented as a Ciruclar Doubly Linked List.