part2: execution time and structure size changes
The changes were related to:

Execution time calculation: previously, it was being calculated using the clock() function from time.h. Now, it is using the timeval structure and the gettimeofday() function.

show_size_of_structures(): the calculation of the sizes of both structures were not correct.
Array: the sizeof must be (int64_t) instead of (myArray)
List: the sizeof must be (struct node) instead of (first)