part3: fix _new_read_from_file()
There was a problem with the method _new_read_from_file(), where in some specific cases an empty line was given while using the method strtok. The solution was to use the methods strdup() and strsep() isntead, and check if the returned value was not NULL and its length is higher than 0.